backend/requirements.txt
tqdm
numpy-stl
gradio_client
scipy
plotly
fastapi
uvicorn
python-multipart
